requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.1
attrs==25.4.0
bcrypt==4.0.1
blis==1.3.3
catalogue==2.0.10
certifi==2026.1.4
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cymem==2.0.13
datasets==4.5.0
dill==0.4.0
ecdsa==0.19.1
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
fastapi==0.128.8
filelock==3.24.2
frozenlist==1.8.0
fsspec==2025.10.0
greenlet==3.3.1
google-auth==2.41.1
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.4.1
idna==3.11
Jinja2==3.1.6
joblib==1.5.3
markdown-it-py==4.0.0
MarkupSafe==3.0.3
mdurl==0.1.2
multidict==6.7.1
multiprocess==0.70.18
murmurhash==1.0.15
numpy==2.4.2
packaging==26.0
pandas==3.0.0
passlib==1.7.4
preshed==3.0.12
propcache==0.4.1
psycopg2-binary==2.9.11
pyarrow==23.0.1
pyasn1==0.6.2
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
PyMySQL==1.1.2
python-dateutil==2.9.0.post0
python-jose==3.5.0
python-multipart==0.0.22
PyYAML==6.0.3
requests==2.32.5
rich==14.3.2
rsa==4.9.1
scikit-learn==1.8.0
scipy==1.17.0
shellingham==1.5.4
six==1.17.0
smart_open==7.5.0
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spacy-lookups-data==1.0.5
SQLAlchemy==2.0.46
srsly==2.5.2
starlette==0.52.1
thinc==8.3.10
threadpoolctl==3.6.0
tqdm==4.67.3
typer==0.23.0
typer-slim==0.23.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
urllib3==2.6.3
uvicorn==0.40.0
wasabi==1.1.3
weasel==0.4.3
wrapt==2.1.1
xxhash==3.6.0
yarl==1.22.0
